Code Update

+++ ktb/2model/src/code/primes.cu
@@ -102,7 +102,7 @@ kernelLauncher(gpu_id);
 // Global Variables.
 long long int pl_end_number = 1000;
 long long int total_primes=0;
-long number_of_gpus = 1;
+int number_of_gpus = 1;
 PrimeHeader pheader;
 GpuHandler gpu_data;
 //long long int end_val = 1000000;
